+++ plugins/Annotations/Annotations.php
@@ -16,9 +16,9 @@ namespace Piwik\Plugins\Annotations;
 class Annotations extends \Piwik\Plugin
 {
     /**
-     * @see Piwik\Plugin::getListHooksRegistered
+     * @see Piwik\Plugin::registerEvents
      */
-    public function getListHooksRegistered()
+    public function registerEvents()
     {
         return array(
             'AssetManager.getStylesheetFiles' => 'getStylesheetFiles',