+++ plugins/CustomVariables/CustomVariables.php
@@ -16,9 +16,9 @@ use Piwik\Tracker;
 class CustomVariables extends \Piwik\Plugin
 {
     /**
-     * @see Piwik\Plugin::getListHooksRegistered
+     * @see Piwik\Plugin::registerEvents
      */
-    public function getListHooksRegistered()
+    public function registerEvents()
     {
         return array(
             'API.getSegmentDimensionMetadata' => 'getSegmentsMetadata',