+++ plugins/Dashboard/Dashboard.php
@@ -18,9 +18,9 @@ use Piwik\WidgetsList;
 class Dashboard extends \Piwik\Plugin
 {
     /**
-     * @see Piwik\Plugin::getListHooksRegistered
+     * @see Piwik\Plugin::registerEvents
      */
-    public function getListHooksRegistered()
+    public function registerEvents()
     {
         return array(
             'AssetManager.getJavaScriptFiles'        => 'getJsFiles',