+++ plugins/PrivacyManager/PrivacyManager.php
@@ -130,9 +130,9 @@ class PrivacyManager extends Plugin
     }
 
     /**
-     * @see Piwik\Plugin::getListHooksRegistered
+     * @see Piwik\Plugin::registerEvents
      */
-    public function getListHooksRegistered()
+    public function registerEvents()
     {
         return array(
             'AssetManager.getJavaScriptFiles'         => 'getJsFiles',