+++ plugins/Transitions/Transitions.php
@@ -14,9 +14,9 @@ namespace Piwik\Plugins\Transitions;
 class Transitions extends \Piwik\Plugin
 {
     /**
-     * @see Piwik\Plugin::getListHooksRegistered
+     * @see Piwik\Plugin::registerEvents
      */
-    public function getListHooksRegistered()
+    public function registerEvents()
     {
         return array(
             'AssetManager.getStylesheetFiles'        => 'getStylesheetFiles',