+++ plugins/VisitsSummary/VisitsSummary.php
@@ -21,9 +21,9 @@ use Piwik\Plugins\VisitsSummary\Reports\Get;
 class VisitsSummary extends \Piwik\Plugin
 {
     /**
-     * @see Piwik\Plugin::getListHooksRegistered
+     * @see Piwik\Plugin::registerEvents
      */
-    public function getListHooksRegistered()
+    public function registerEvents()
     {
         return array(
             'AssetManager.getStylesheetFiles' => 'getStylesheetFiles',