+++ tests/PHPUnit/Unit/DeprecatedMethodsTest.php
@@ -69,6 +69,7 @@ class DeprecatedMethodsTest extends \PHPUnit_Framework_TestCase
 
         $this->assertDeprecatedMethodIsRemovedInPiwik3('\Piwik\Menu\MenuAbstract', 'add');
         $this->assertDeprecatedMethodIsRemovedInPiwik3('\Piwik\Archive', 'getDataTableFromArchive');
+        $this->assertDeprecatedMethodIsRemovedInPiwik3('\Piwik\Plugin', 'getListHooksRegistered');
     }
 
     private function assertDeprecatedMethodIsRemoved($className, $method, $removalDate)